Make key bindings potentially platform specific
This includes the ability for the key bindings in the project file to
specify a `platform` that they should be for, allowing a project to
contain different key bindings depending on the platform currently in
use.

This makes sure that if the same project is used across multiple
platforms that the bindings can be customized to platform defaults.