Exclude a file and update the README
Much like we excluded the workspace file from the package, this also
excludes the sublime project file. This addition will ensure that the
zip file that PackageControl downloads from GitHub won't have files
in it that are not needed.

In addition, the README has been changed slightly to account for the
face that the changes it called out have already been made.