Include a command to easily edit project keys
Include a simple command that will use the internal `edit_settings`
command to open a key bindings window with the platform defaults on the
left and the project on the right.

In addition, there is now a menu entry and a command palette entry for
triggering the command, making the addition of key bindings a little
bit easier.